requirements_Version2 (1).txt
streamlit
Pillow
opencv-python-headless
numpy
cairosvg
pdf2image